factored out well overview layout in separate file
