+++ webviz_subsurface/plugins/_well_analysis/_layout/__init__.py
@@ -1,1 +1,3 @@
-from .layout import main_layout, LayoutElements
+from .main_layout import main_layout
+from .well_control_layout import WellControlLayoutElements
+from .well_overview_layout import WellOverviewLayoutElements
